+++ internal/admission/validator_test.go
@@ -94,7 +94,20 @@ func (f fakeServicesProvider) GetRoutesService() (kong.AbstractRouteService, boo
 }
 
 func TestKongHTTPValidator_ValidatePlugin(t *testing.T) {
-	store, _ := store.NewFakeStore(store.FakeObjects{})
+	store, _ := store.NewFakeStore(store.FakeObjects{
+		Secrets: []*corev1.Secret{
+			{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "",
+					Name:      "conf-secret",
+				},
+				Data: map[string][]byte{
+					"valid-conf":   []byte(`{"foo":"bar"}`),
+					"invalid-conf": []byte(`{"foo":"baz}`),
+				},
+			},
+		},
+	})
 	type args struct {
 		plugin kongv1.KongPlugin
 	}
@@ -139,7 +152,59 @@ func TestKongHTTPValidator_ValidatePlugin(t *testing.T) {
 			},
 			wantOK:      false,
 			wantMessage: ErrTextPluginConfigInvalid,
-			wantErr:     true,
+			wantErr:     false,
+		},
+		{
+			name:      "plugin has valid configPatches",
+			PluginSvc: &fakePluginSvc{valid: true},
+			args: args{
+				plugin: kongv1.KongPlugin{
+					PluginName: "key-auth",
+					Config: apiextensionsv1.JSON{
+						Raw: []byte(`{"k1":"v1"}`),
+					},
+					ConfigPatches: []kongv1.ConfigPatch{
+						{
+							Path: "/foo",
+							ValueFrom: kongv1.ConfigSource{
+								SecretValue: kongv1.SecretValueFromSource{
+									Secret: "conf-secret",
+									Key:    "valid-conf",
+								},
+							},
+						},
+					},
+				},
+			},
+			wantOK:      true,
+			wantMessage: "",
+			wantErr:     false,
+		},
+		{
+			name:      "plugin has invalid configPatches",
+			PluginSvc: &fakePluginSvc{},
+			args: args{
+				plugin: kongv1.KongPlugin{
+					PluginName: "key-auth",
+					Config: apiextensionsv1.JSON{
+						Raw: []byte(`{"k1":"v1"}`),
+					},
+					ConfigPatches: []kongv1.ConfigPatch{
+						{
+							Path: "/foo",
+							ValueFrom: kongv1.ConfigSource{
+								SecretValue: kongv1.SecretValueFromSource{
+									Secret: "conf-secret",
+									Key:    "invalid-conf",
+								},
+							},
+						},
+					},
+				},
+			},
+			wantOK:      false,
+			wantMessage: ErrTextPluginConfigInvalid,
+			wantErr:     false,
 		},
 		{
 			name:      "plugin ConfigFrom references non-existent Secret",
@@ -157,7 +222,7 @@ func TestKongHTTPValidator_ValidatePlugin(t *testing.T) {
 			},
 			wantOK:      false,
 			wantMessage: ErrTextPluginSecretConfigUnretrievable,
-			wantErr:     true,
+			wantErr:     false,
 		},
 		{
 			name:      "failed to retrieve validation info",
@@ -179,23 +244,40 @@ func TestKongHTTPValidator_ValidatePlugin(t *testing.T) {
 				},
 				ingressClassMatcher: fakeClassMatcher,
 			}
-			got, got1, err := validator.ValidatePlugin(context.Background(), tt.args.plugin)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("KongHTTPValidator.ValidatePlugin() error = %v, wantErr %v", err, tt.wantErr)
-				return
+			gotOK, gotMessage, err := validator.ValidatePlugin(context.Background(), tt.args.plugin)
+			assert.Equalf(t, tt.wantOK, gotOK,
+				"KongHTTPValidator.ValidatePlugin() want OK: %v, got OK: %v",
+				tt.wantOK, gotOK,
+			)
+			if tt.wantMessage != "" {
+				assert.Containsf(t, gotMessage, tt.wantMessage,
+					"KongHTTPValidator.ValidatePlugin() want message: %v, got message: %v",
+					tt.wantMessage, gotMessage,
+				)
 			}
-			if got != tt.wantOK {
-				t.Errorf("KongHTTPValidator.ValidatePlugin() got = %v, want %v", got, tt.wantOK)
-			}
-			if got1 != tt.wantMessage {
-				t.Errorf("KongHTTPValidator.ValidatePlugin() got1 = %v, want %v", got1, tt.wantMessage)
-			}
+			assert.Equalf(t, tt.wantErr, err != nil,
+				"KongHTTPValidator.ValidatePlugin() wantErr %v, got error %v",
+				tt.wantErr, err,
+			)
 		})
 	}
 }
 
 func TestKongHTTPValidator_ValidateClusterPlugin(t *testing.T) {
-	store, _ := store.NewFakeStore(store.FakeObjects{})
+	store, _ := store.NewFakeStore(store.FakeObjects{
+		Secrets: []*corev1.Secret{
+			{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "default",
+					Name:      "conf-secret",
+				},
+				Data: map[string][]byte{
+					"valid-conf":   []byte(`{"foo":"bar"}`),
+					"invalid-conf": []byte(`{"foo":"baz}`),
+				},
+			},
+		},
+	})
 	type args struct {
 		plugin kongv1.KongClusterPlugin
 	}
@@ -240,7 +322,61 @@ func TestKongHTTPValidator_ValidateClusterPlugin(t *testing.T) {
 			},
 			wantOK:      false,
 			wantMessage: ErrTextPluginConfigInvalid,
-			wantErr:     true,
+			wantErr:     false,
+		},
+		{
+			name:      "plugin has valid configPatches",
+			PluginSvc: &fakePluginSvc{valid: true},
+			args: args{
+				plugin: kongv1.KongClusterPlugin{
+					PluginName: "key-auth",
+					Config: apiextensionsv1.JSON{
+						Raw: []byte(`{"k1":"v1"}`),
+					},
+					ConfigPatches: []kongv1.NamespacedConfigPatch{
+						{
+							Path: "/foo",
+							ValueFrom: kongv1.NamespacedConfigSource{
+								SecretValue: kongv1.NamespacedSecretValueFromSource{
+									Namespace: "default",
+									Secret:    "conf-secret",
+									Key:       "valid-conf",
+								},
+							},
+						},
+					},
+				},
+			},
+			wantOK:      true,
+			wantMessage: "",
+			wantErr:     false,
+		},
+		{
+			name:      "plugin has invalid configPatches",
+			PluginSvc: &fakePluginSvc{},
+			args: args{
+				plugin: kongv1.KongClusterPlugin{
+					PluginName: "key-auth",
+					Config: apiextensionsv1.JSON{
+						Raw: []byte(`{"k1":"v1"}`),
+					},
+					ConfigPatches: []kongv1.NamespacedConfigPatch{
+						{
+							Path: "/foo",
+							ValueFrom: kongv1.NamespacedConfigSource{
+								SecretValue: kongv1.NamespacedSecretValueFromSource{
+									Namespace: "default",
+									Secret:    "conf-secret",
+									Key:       "invalid-conf",
+								},
+							},
+						},
+					},
+				},
+			},
+			wantOK:      false,
+			wantMessage: ErrTextPluginConfigInvalid,
+			wantErr:     false,
 		},
 		{
 			name:      "plugin ConfigFrom references non-existent Secret",
@@ -259,7 +395,7 @@ func TestKongHTTPValidator_ValidateClusterPlugin(t *testing.T) {
 			},
 			wantOK:      false,
 			wantMessage: ErrTextPluginSecretConfigUnretrievable,
-			wantErr:     true,
+			wantErr:     false,
 		},
 		{
 			name:      "failed to retrieve validation info",
@@ -291,17 +427,22 @@ func TestKongHTTPValidator_ValidateClusterPlugin(t *testing.T) {
 				},
 				ingressClassMatcher: fakeClassMatcher,
 			}
-			got, got1, err := validator.ValidateClusterPlugin(context.Background(), tt.args.plugin)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("KongHTTPValidator.ValidatePlugin() error = %v, wantErr %v", err, tt.wantErr)
-				return
+
+			gotOK, gotMessage, err := validator.ValidateClusterPlugin(context.Background(), tt.args.plugin)
+			assert.Equalf(t, tt.wantOK, gotOK,
+				"KongHTTPValidator.ValidateClusterPlugin() want OK: %v, got OK: %v",
+				tt.wantOK, gotOK,
+			)
+			if tt.wantMessage != "" {
+				assert.Containsf(t, gotMessage, tt.wantMessage,
+					"KongHTTPValidator.ValidateClusterPlugin() want message: %v, got message: %v",
+					tt.wantMessage, gotMessage,
+				)
 			}
-			if got != tt.wantOK {
-				t.Errorf("KongHTTPValidator.ValidatePlugin() got = %v, want %v", got, tt.wantOK)
-			}
-			if got1 != tt.wantMessage {
-				t.Errorf("KongHTTPValidator.ValidatePlugin() got1 = %v, want %v", got1, tt.wantMessage)
-			}
+			assert.Equalf(t, tt.wantErr, err != nil,
+				"KongHTTPValidator.ValidateClusterPlugin() wantErr %v, got error %v",
+				tt.wantErr, err,
+			)
 		})
 	}
 }